move same padding pair to utils.py
